refactor(admin/filters): align admin filter components with new filter hooks and options
